+++ src/main/java/com/example/cateam4spring/service/StudentService.java
@@ -7,7 +7,7 @@ public interface StudentService {
 	//public boolean Authenticate(Student Student);
 	//public Student findStudentById(String id);
 	
-	public Integer findGpaById(Integer id);
+	public Double findGpaById(Integer id);
 	
 	
 	public Double calculateGPA(Integer id);
